further cleanup | #1694
