+++ extension/chrome/elements/reply_pubkey_mismatch.ts
@@ -62,7 +62,7 @@ Catch.try(async () => {
   $('#send_btn').off().click(Ui.event.prevent('double', async target => {
     $(target).text('sending..');
     const body = { 'text/plain': $('#input_text').get(0).innerText };
-    const message = await Api.common.msg(acctEmail, from, to, subject, body, [att], threadId);
+    const message = await Google.createMsgObj(acctEmail, from, to, subject, body, [att], threadId);
     const replyHeaders = await determineReplyHeaders();
     message.headers['In-Reply-To'] = replyHeaders['In-Reply-To'];
     message.headers.References = replyHeaders.References;
